Added deduplication when following deps (see notes
The built-in list(set(...)) trick to de-duplicate lists has been the
approach for any sources() or includes() call to a `Target` when
`ignore_deps` is set to `False`. This is to prevent accidentally picking
up multiple entries of the same shared dependency in common, for
example.

This deduplication is not applied when polling targets in isolation,
however, which is the behaviour seen when `ignore_deps=True`.

The only exception to this is the include_dirs -- as many header files
tend to share the same parent directory structure, it is assumed
de-duplication will frequently be required even on the individual target
level.